Borrow joint_trajectory_action.py from ur_driver
This contains the config for my robot.  I didn't want to edit the .py file in ur_driver since that will make things more difficult to merge.  I need to make a new joint_trajectory_action that behaves like the industrial_robot_client eventually, but this will work for now.