Remove clearCommands() from cmdThread's exception handler
If I restart the PLC application the cmd socket will be lost but it won't realize it until it tries to send a command, so the 1st CommandList after a reconnect would always fail and then clear the cmd list.  Maybe I should add a timer or reconnect limit before clearing?