Use CommandHandler::initialize method
initialize is called by addHandler after setCommandRegister().  This will work better than the constructor for handlers that need info from the register like the number of joints.  I'm undecided if I want to move everything to initialize and make it pure virtual and forget the constructor.  That might make it difficult to support lambdas unless I make another CommandHandler type.