Change commandListCb to be more restrictive on failures
Connector::commandListCb is now much more restrictive.  Failing to find a CommandHandler or failing to processMsg will cause the current commands to be cleared and prevent any of the commands in this list from being added. 

If I make a mandatory Abort CommandHandler then I could abort it on the robot too.  I need to think about how to handle this with multiple robots.  Something like an IO command failing in the middle of a path could make 1 robot stop and the other crash into it.